Renaming mod name to fix clippy warning
