+++ data-node/src/stream_manager/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod fetcher;
 pub(crate) mod range;
 pub(crate) mod stream;
-pub(crate) mod stream_manager;
+pub(crate) mod manager;
 pub(crate) mod window;
 
-pub(crate) use stream_manager::StreamManager;
+pub(crate) use manager::StreamManager;
